debugged some ga subtelties

+++ find_best_attributes.py
@@ -8,7 +8,7 @@ Created on 27/09/2010
 @author: peter
 """
 
-import sys, os, random, math, csv, preprocess_soybeans, weka_classifiers as WC, arff, ga
+import sys, os, random, math, time, csv, preprocess_soybeans, weka_classifiers as WC, arff, ga
 
 # Set random seed so that each run gives same results
 random.seed(555)
@@ -36,18 +36,41 @@ def writeArffForInclusiveSubset(filename, data, attributes, subset):
     arff.writeArff(filename, None, 'find_best_attr', attrs_subset, data_subset)
 
 def getAccuracyForInclusiveSubset(algo_key, data, attributes, subset):
-    training_filename =  makeFileName('find-best-attr', algo_key, None, 'arff')
+    training_filename = makeFileName('find-best-attr', algo_key, None, 'arff')
+    training_filename = 'find-best-attr.arff'
+
+    try:
+        os.remove(training_filename)
+    except:
+        pass
+
     writeArffForInclusiveSubset(training_filename, data, attributes, subset)
     result = WC.getAccuracyAlgoKey(algo_key, class_index, training_filename)
     # print 'result',result
-    for i in range(1000):
-        try:
-            os.remove(training_filename)
-            break
-        except:
-            pass
-        
-    
+    if False:
+        deleted = False
+        for tries in range(10):
+            for i in range(1000):
+                exist = os.path.exists(training_filename)
+                if exist:
+                    print training_filename, 'exists !!!!'
+                    os.remove(training_filename)
+                try:
+                    print '+++', training_filename
+                    if exist:
+                        print training_filename, 'exists'
+                        os.remove(training_filename)
+                    deleted = True
+                    break
+                except:
+                    print '***'
+                    pass
+            if not deleted:
+                print 'cannot delete', training_filename
+                time.sleep(1)
+            else:
+                break
+
     return result
 
 def getAccuracyForSubset0(algo_key, data, attributes, subset):
@@ -113,17 +136,16 @@ def findBestAttributesForSubsetSize(base_filename, algo_key, data, attributes, p
     results = []
     existing_subsets = []
     history_of_best = []
-    best_score = 0.0
-
+   
     print base_filename, algo_key, len(attributes), len(previous_results), subset_size
 
     def addSubset(subset):
-        if not subset in existing_subsets:
-            r = getSubsetResultDict(algo_key, data, attributes, subset)
-            results.append(r)
-            existing_subsets.append(subset)
-            results.sort(key = lambda x: -x['score'])
-            best_score = results[0]['score']
+        if subset:
+            if not subset in existing_subsets:
+                r = getSubsetResultDict(algo_key, data, attributes, subset)
+                results.append(r)
+                existing_subsets.append(subset)
+                results.sort(key = lambda x: -x['score'])
 
     for s in previous_results:
         assert(len(s['subset']) == subset_size - 1)
@@ -135,51 +157,95 @@ def findBestAttributesForSubsetSize(base_filename, algo_key, data, attributes, p
     done = False
     for r in previous_best:
         for i in range(len(attributes)):
-            s = sorted(set(r['subset'][:] + [getRandomExcluding(num_attrs, r['subset'] + [class_index])]))
-            assert(len(s) == subset_size)
-            if not s in subsets:
-                subsets.append(s)
-                if len(subsets) >= candidates_per_round:
-                    done = True
-                    break
+            if not i in r['subset'] and 1 != class_index:
+                s = sorted(set(r['subset'][:] + [i]))
+            #s = sorted(set(r['subset'][:] + [getRandomExcluding(num_attrs, r['subset'] + [class_index])]))
+                assert(len(s) == subset_size)
+                if not s in subsets:
+                    subsets.append(s)
+                    if len(subsets) >= candidates_per_round:
+                        done = True
+                        break
         if done:
             break
 
     # Compute scores for initial population
+    superset = []
     for subset in subsets:
         addSubset(subset)
+        superset = list(set(superset + subset))
+    print 'superset', len(superset), sorted(superset)
+    assert(len(superset) == len(attributes))
+    
+    print 'subsets', subsets
 
-    for cnt in range(1000):
-        found = False
-        for j in range(10):
-            #print 'len(results)', len(results)
-            i1,i2 = ga.spinRouletteWheelTwice(results)
-            c1,c2 = ga.crossOver(results[i1]['subset'], results[i2]['subset'])
-            if not c1 in existing_subsets and not c2 in existing_subsets:
-                found = True
-                break
-        if not found:
-            print '1. Converged after', cnt, 'GA rounds'
-            break
-        addSubset(c1)
-        addSubset(c2)
-
-        # Test for convergence. Top <convergence_number> scores are the sme
-        convergence_number = 10
-        print ['%.1f%%' % x['score'] for x in results[:10]]
-
-        history_of_best.append(best_score)
-        history_of_best.sort(key = lambda x: -x)
-        if len(history_of_best) >= convergence_number:
-            print 'history_of_best =', history_of_best[:10]
-            converged = True
-            for i in range(1, convergence_number):
-                if history_of_best[i] != history_of_best[0]:
-                    converged = False
+    if subset_size > 1:
+        
+        counters = [0]*20
+        for cnt in range(1000):
+            found = False
+            for j in range(1000):
+                #print 'len(results)', len(results)
+                i1,i2 = ga.spinRouletteWheelTwice(results)
+                c1,c2 = ga.crossOver(results[i1]['subset'], results[i2]['subset'])
+                print '*', i1,i2, results[i1]['subset'], results[i2]['subset'], '=>', c1, c2
+                if i1 < len(counters):
+                    counters[i1] += 1
+                if i2 < len(counters):
+                    counters[i2] += 1
+                
+                if c1 in existing_subsets:
+                    c1 = None
+                if c2 in existing_subsets:
+                    c2 = None
+                if c1 or c2:
+                    found = True
+                    print '***', c1
+                    print '+++', c2
                     break
-            if converged:
-                print '2. Converged after', cnt, 'GA rounds'
+                
+            if not found:
+                c1 = None 
+                c2 = None
+                for j in range(1000):
+                    i1,i2 = ga.spinRouletteWheelTwice(results)
+                    c1 = list(set(results[i1]['subset'] + results[i2]['subset']))
+                    random.shuffle(c1)
+                    c1 = sorted(c1[:subset_size])
+                    if not c1 in existing_subsets:
+                        found = True
+                        print '***', c1
+                        break
+                    
+            if not found:
+                print '1. Converged after', cnt, 'GA rounds'
+                print existing_subsets
+                print 'counters', counters
+                for i,n in enumerate(counters):
+                    print '%2d: %4d' % (i,n), results[i]['subset'], results[i]['score']
+                exit()
                 break
+            
+            addSubset(c1)
+            addSubset(c2)
+    
+            # Test for convergence. Top <convergence_number> scores are the sme
+            convergence_number = 10
+            print ['%.1f%%' % x['score'] for x in results[:10]]
+    
+            history_of_best.append(results[0]['score'])
+            print 'history_of_best:', results[0]['score'], '=>', history_of_best[:10]
+            history_of_best.sort(key = lambda x: -x)
+            if len(history_of_best) >= convergence_number:
+                print 'history_of_best =', history_of_best[:10]
+                converged = True
+                for i in range(1, convergence_number):
+                    if history_of_best[i] != history_of_best[0]:
+                        converged = False
+                        break
+                if converged:
+                    print '2**. Converged after', cnt, 'GA rounds'
+                    break
 
     best_arff = makeFileName(base_filename, algo_key, subset_size, 'arff')
     best_results = makeFileName(base_filename, algo_key, subset_size, 'results')
@@ -244,9 +310,27 @@ if __name__ == '__main__':
     print 'Algos to test 2:', WC.all_algo_keys
     for k in WC.all_algo_keys:
         assert(k in WC.algo_dict.keys())
-    assert('does not exist' not in WC.algo_dict.keys())
-  
-   # algo_key = 'BayesNet' 
+    assert('does not exist' not in WC.algo_dict.keys()) 
+    
+    log_file = file('log.txt', 'w+')
+    log_file.write('Opening ----------------\n')
+    log_file.flush()
+    
+    time.sleep(1)
+   
     for algo_key in WC.all_algo_keys:
-        findBestAttributes(filename, algo_key, data, attributes)
-
+       findBestAttributes(filename, algo_key, data, attributes)
+       
+    if False:
+        for algo_key in WC.all_algo_keys:
+            try:
+                findBestAttributes(filename, algo_key, data, attributes)
+            except:
+                try:
+                    print 'findBestAttributes(%s, %s) failed' % (filename, algo_key)
+                    log_file.write('findBestAttributes(%s, %s) failed\n' % (filename, algo_key))
+                    log_file.flush()
+                except:
+                    pass
+                
+
